% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formality.R
\name{plot.Formality}
\alias{plot.Formality}
\title{Plots a Formality Object}
\usage{
\method{plot}{Formality}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{The Formality object}

\item{plot}{logical.  If \code{TRUE} the output is plotted.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of the three \pkg{ggplot2} objects that make the
combined plot.
}
\description{
Plots a Formality object.
}
