% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formality.R
\name{formality}
\alias{formality}
\title{Formality Score}
\usage{
formality(text.var, grouping.var = NULL, order.by.formality = TRUE, ...)
}
\arguments{
\item{text.var}{The text string variable or another \code{formality} object.
The latter recycles the part of speech tagging and is much faster.}

\item{grouping.var}{The grouping variable(s).  Default \code{NULL} generates
one word list for all text.  Also takes a single grouping variable or a list
of 1 or more grouping variables.  If \code{TRUE} an \code{id} variable is
used with a \code{seq_along} the \code{text.var}.}

\item{order.by.formality}{logical.  If \code{TRUE} orders the results
descending by formality score.}

\item{\ldots}{ignored.}
}
\description{
Formality score by grouping variable(s).
}
\details{
Heylighen & Dewaele(2002)'s formality \emph{F-measure} is defined
formally as:

\deqn{F = 50(\frac{n_{f}-n_{c}}{N} + 1)}{F = 50(n_f-n_c/N + 1)}

Where:

\deqn{f = \left \{noun, \;adjective, \;preposition, \;article\right \}}{f = {noun,adjective, preposition, article}}
\deqn{c = \left \{pronoun, \;verb, \;adverb, \;interjection\right \}}{c = {pronoun, verb, adverb, interjection}}
\deqn{N = \sum{(n_{f}\;+ \;n_{c})}}{N = \sum(f + c)}

This yields an \emph{F-measure} between \eqn{0} and \eqn{100}\%, with
completely contextualizes language on the zero end and completely formal
language on the \eqn{100} end.
}
\section{Warning}{
 Heylighen & Dewaele (2002) state, "At present, a sample would
probably need to contain a few hundred words for the measure to be minimally
reliable. For single sentences, the F-value should only be computed for
purposes of illustration" (p. 24).
}
\examples{
data(presidential_debates_2012)
(form1 <- with(presidential_debates_2012, formality(dialogue, person)))
with(presidential_debates_2012, formality(form1, list(person, time))) #recycle form 1 for speed
}
\references{
Heylighen, F. (1999). Advantages and limitations of formal expression. doi:10.1023/A:1009686703349 \cr \cr
Heylighen, F. & Dewaele, J.-M. (1999). Formality of language: Definition, measurement and behavioral determinants. Center "Leo Apostel", Free University of Brussels. Retrieved from \url{http://pespmc1.vub.ac.be/Papers/Formality.pdf} \cr \cr
Heylighen, F. & Dewaele, J.-M. (2002). Variation in the contextuality of language: An empirical measure. \emph{Foundations of Science}, 7(3), 293-340. doi:10.1023/A:1019661126744
}
\keyword{formality}

